# Support all dtypes in every sorting function in `cupy.cuda.thrust`