--- a/cupy/cuda/common.pxd
+++ b/cupy/cuda/common.pxd
@@ -1,5 +1,7 @@
 # distutils: language = c++
 
+from libcpp cimport bool as bool_t
+
 cdef extern from '../cuda/cupy_common.h':  # thru parent to import in core
     ctypedef char cpy_byte
     ctypedef unsigned char cpy_ubyte
@@ -13,4 +15,5 @@ cdef extern from '../cuda/cupy_common.h':  # thru parent to import in core
     ctypedef double cpy_double
     ctypedef struct cpy_complex64 'cuComplex'
     ctypedef struct cpy_complex128 'cuDoubleComplex'
+    ctypedef bool_t cpy_bool 'bool'
     ctypedef struct cpy_half '__half'