--- a/cupy/cuda/cupy_common.h
+++ b/cupy/cuda/cupy_common.h
@@ -15,14 +15,15 @@ typedef float cpy_float;
 typedef double cpy_double;
 typedef cuComplex cpy_complex64;
 typedef cuDoubleComplex cpy_complex128;
+typedef bool cpy_bool;
 
 #if (__CUDACC_VER_MAJOR__ > 9 || (__CUDACC_VER_MAJOR__ == 9 && __CUDACC_VER_MINOR__ == 2)) \
     && (__CUDA_ARCH__ >= 530 || !defined(__CUDA_ARCH__))
 #include <cuda_fp16.h>
-typedef __half cpy_half;
 #else
 // still declare it to avoid build time errors, but it won't be usable
-typedef struct __half cpy_half;
+struct __half;
 #endif
+typedef __half cpy_half;
 
 #endif // INCLUDE_GUARD_CUPY_CUDA_COMMON_H