--- a/cupy/cuda/cupy_common.h
+++ b/cupy/cuda/cupy_common.h
@@ -2,6 +2,11 @@
 #define INCLUDE_GUARD_CUPY_COMMON_H
 
 #include "cupy_cuComplex.h"
+/*
+   cuda_fp16.h should only be included in C++ codes that actually
+   need __half. For bookkeeping purposes, in Cython codes we simply
+   use cupy.float16.
+*/
 
 typedef char cpy_byte;
 typedef unsigned char cpy_ubyte;
@@ -17,13 +22,5 @@ typedef cuComplex cpy_complex64;
 typedef cuDoubleComplex cpy_complex128;
 typedef bool cpy_bool;
 
-#ifndef CUPY_NO_CUDA
-#if (__CUDACC_VER_MAJOR__ > 9 || (__CUDACC_VER_MAJOR__ == 9 && __CUDACC_VER_MINOR__ == 2)) \
-    && (__CUDA_ARCH__ >= 530 || !defined(__CUDA_ARCH__))
-#include <cuda_fp16.h>
-#endif
-#endif // CUPY_NO_CUDA
-
-typedef struct __half cpy_half;
 
 #endif // INCLUDE_GUARD_CUPY_COMMON_H