… cupy/cuda/cupy_thrust.h
… cupy/cuda/thrust.pyx
… tests/cupy_tests/sorting_tests/test_sort.py