All kinds of sort functions work with all dtypes now!

~~Ongoing, **DO NOT MERGE!**~~
- ~~msort~~
- ~~lexsort~~
- ~~remove redundant tests~~
- ~~clean up~~
- ~~test for performance regression~~ LGTM