@leofang 2020-04-15
Will think about how to fix Travis tomorrow...😞 Likely I forgot a `CUPY_NO_CUDA` somewhere...