@leofang 2020-04-16
@takagi e673795 fixed the build error with or without `CUPY_NO_CUDA`. I will start looking into possible performance implications. In the meanwhile, could you kick off the CI to see if I introduced bugs? (I only tested `test_sort.py` locally.) Thanks!