@leofang 2020-04-16
> I am amazed it managed to compile until now, lol

If you meant the bug caused by the repetitive guard name, yeah it's surprising until you figure out `cupy_common.h` and `common.pxd` is only included in the thrust module...

One thing we could do next, if `cupy_common.h` is meant for type aliasing, is to move this part to it: https://github.com/cupy/cupy/blob/9d8dca77507e47bee62a2ee0ba1da4f67874e3be/cupy/cuda/cupy_cub.h#L4-L16
This will make the alias more reusable across modules.

OK I also need to fix Appveyor now 😞 I suspect we should not specialize for `__half` on host...Will wait for the CI to finish and push a test commit.