@leofang 2020-04-16
before:
- compile time:
real	1m48.061s
user	1m44.241s
sys	0m3.883s
(`cupy/core/_routines_sorting.pyx` and `cupy/cuda/thrust.pyx` only; set `CUPY_NVCC_GENERATE_CODE=arch=compute_75,code=sm_75`)
- performance:
```python
>>> import cupy as cp
>>> from cupyx.time import repeat
>>>
>>> # 1M elements
>>> def f():
...     a = cp.random.random((1000000,))
...     a.sort()
... 
>>> print(repeat(f, ()))
f                   :    CPU:   91.497 us   +/- 5.006 (min:   87.961 / max:  223.452) us     GPU:  961.748 us   +/-20.848 (min:  948.576 / max: 1155.328) us
>>> 
>>> # 10M elements
>>> def f():
...     a = cp.random.random((10000000,))
...     a.sort()
...     del a
... 
>>> print(repeat(f, (), n=100))
f                   :    CPU:   94.498 us   +/- 4.014 (min:   90.030 / max:  112.589) us     GPU: 6645.967 us   +/- 7.638 (min: 6629.408 / max: 6671.296) us
```

This PR:
- compile time: 
real	2m4.182s
user	2m0.274s
sys	0m3.941s
- performance:
```python
>>> import cupy as cp
>>> from cupyx.time import repeat
>>>
>>> # 1M elements
>>> def f():
...     a = cp.random.random((1000000,))
...     a.sort()
... 
>>> print(repeat(f, ()))
f                   :    CPU:   92.628 us   +/- 4.465 (min:   88.703 / max:  208.141) us     GPU:  965.287 us   +/-18.920 (min:  951.872 / max: 1152.832) us
>>>
>>> # 10M elements
>>> def f():
...     a = cp.random.random((10000000,))
...     a.sort()
...     del a
... 
>>> print(repeat(f, (), n=100))
f                   :    CPU:   94.090 us   +/- 5.719 (min:   89.509 / max:  121.101) us     GPU: 6631.661 us   +/- 8.929 (min: 6612.512 / max: 6660.576) us
```