@leofang 2020-04-16
@emcastillo @takagi I feel the right way to fix the compatibility with older CUDA and to increase reusability is to avoid exposing `__half` to the Cython space. I think it can be done by moving this part
https://github.com/cupy/cupy/blob/e339e1f5513afc0922b406e6e721c86d77d42a23/cupy/cuda/cupy_cub.cu#L407-L436
to `cupy_common.h`, and do the type dispatch in the C++ space. May I proceed with this?