@leofang 2020-04-16
...I feel I might have missed something obvious? How is fp16 supported in CuPy?