@leofang 2020-04-17
I worked around the issue of exposing `__half` to Cython, as I don't wanna do refactoring in this PR 😅 Also, I added a few more tests with the presence of NaN, which should have been added in my earlier PR #2745.